# frontend/openapi-schema.json
{
  "components": {
    "schemas": {
...
      "ApiResponse_DocumentResponse_": {
        "properties": {
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DocumentResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          },
          "meta": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ResponseMeta"
              },
              {
                "type": "null"
              }
            ]
          },
          "success": {
            "default": true,
            "title": "Success",
            "type": "boolean"
          }
        },
        "title": "ApiResponse[DocumentResponse]",
        "type": "object"
      },
...
      "ApiResponse_IndexBuildListResponse_": {
        "properties": {
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/IndexBuildListResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          },
          "meta": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ResponseMeta"
              },
              {
                "type": "null"
              }
            ]
          },
          "success": {
            "default": true,
            "title": "Success",
            "type": "boolean"
          }
        },
        "title": "ApiResponse[IndexBuildListResponse]",
        "type": "object"
      },
      "ApiResponse_IndexBuildResponse_": {
        "properties": {
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/IndexBuildResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          },
          "meta": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ResponseMeta"
              },
              {
                "type": "null"
              }
            ]
          },
          "success": {
            "default": true,
            "title": "Success",
            "type": "boolean"
          }
        },
        "title": "ApiResponse[IndexBuildResponse]",
        "type": "object"
      },
...
      "ApiResponse_LifecycleJobResponse_": {
        "properties": {
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/LifecycleJobResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          },
          "meta": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ResponseMeta"
              },
              {
                "type": "null"
              }
            ]
          },
          "success": {
            "default": true,
            "title": "Success",
            "type": "boolean"
          }
        },
        "title": "ApiResponse[LifecycleJobResponse]",
        "type": "object"
      },
...
      "DocumentResponse": {
        "description": "Serialized document metadata returned by the API.",
        "properties": {
          "accepted_parser": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Accepted Parser"
          },
          "content_sha256": {
            "title": "Content Sha256",
            "type": "string"
          },
          "content_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Content Type"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "deleted_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Deleted At"
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          },
          "extraction_method": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Extraction Method"
          },
          "filename": {
            "title": "Filename",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "job_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Job Id"
          },
          "language": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Language"
          },
          "ocr_lang": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ocr Lang"
          },
          "page_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Page Count"
          },
          "parse_quality_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parse Quality Score"
          },
          "parsed_text_storage_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parsed Text Storage Key"
          },
          "parser_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parser Name"
          },
          "parser_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parser Version"
          },
          "project_id": {
            "format": "uuid",
            "title": "Project Id",
            "type": "string"
          },
          "size_bytes": {
            "title": "Size Bytes",
            "type": "integer"
          },
          "status": {
            "$ref": "#/components/schemas/DocumentStatus"
          },
          "storage_key": {
            "title": "Storage Key",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "project_id",
          "filename",
          "content_type",
          "size_bytes",
          "storage_key",
          "content_sha256",
          "status",
          "version",
          "deleted_at",
          "created_at",
          "updated_at"
        ],
        "title": "DocumentResponse",
        "type": "object"
      },
      "DocumentStatus": {
        "description": "Ingestion lifecycle \u2014 full enum defined upfront; phases activate subsets.",
        "enum": [
          "uploaded",
          "queued",
          "parsing",
          "chunking",
          "chunked",
          "failed",
          "embedding",
          "embedded",
          "indexing",
          "ready",
          "deleting",
          "purging"
        ],
        "title": "DocumentStatus",
        "type": "string"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "IndexBuildListResponse": {
        "properties": {
          "active_build_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Active Build Id"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/IndexBuildResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "previous_build_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Previous Build Id"
          }
        },
        "required": [
          "items",
          "active_build_id",
          "previous_build_id"
        ],
        "title": "IndexBuildListResponse",
        "type": "object"
      },
      "IndexBuildOperation": {
        "enum": [
          "ingest",
          "reprocess",
          "reembed",
          "reindex",
          "delete",
          "purge",
          "migration"
        ],
        "title": "IndexBuildOperation",
        "type": "string"
      },
      "IndexBuildResponse": {
        "properties": {
          "activated_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Activated At"
          },
          "chunk_count": {
            "title": "Chunk Count",
            "type": "integer"
          },
          "configuration_hash": {
            "title": "Configuration Hash",
            "type": "string"
          },
          "corpus_fingerprint": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Corpus Fingerprint"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "document_count": {
            "title": "Document Count",
            "type": "integer"
          },
          "embedding_set_version": {
            "title": "Embedding Set Version",
            "type": "integer"
          },
          "failure_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Failure Code"
          },
          "failure_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Failure Message"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "job_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Job Id"
          },
          "keyword_count": {
            "title": "Keyword Count",
            "type": "integer"
          },
          "manifest": {
            "additionalProperties": true,
            "title": "Manifest",
            "type": "object"
          },
          "operation": {
            "$ref": "#/components/schemas/IndexBuildOperation"
          },
          "project_id": {
            "format": "uuid",
            "title": "Project Id",
            "type": "string"
          },
          "state": {
            "$ref": "#/components/schemas/IndexBuildState"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "validated_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Validated At"
          },
          "vector_count": {
            "title": "Vector Count",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "project_id",
          "job_id",
          "state",
          "operation",
          "embedding_set_version",
          "configuration_hash",
          "corpus_fingerprint",
          "document_count",
          "chunk_count",
          "vector_count",
          "keyword_count",
          "manifest",
          "validated_at",
          "activated_at",
          "failure_code",
          "failure_message",
          "created_at",
          "updated_at"
        ],
        "title": "IndexBuildResponse",
        "type": "object"
      },
      "IndexBuildState": {
        "enum": [
          "building",
          "validated",
          "active",
          "retained",
          "failed",
          "superseded"
        ],
        "title": "IndexBuildState",
        "type": "string"
      },
...
      "LifecycleJobResponse": {
        "properties": {
          "build_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Build Id"
          },
          "created": {
            "title": "Created",
            "type": "boolean"
          },
          "job_id": {
            "format": "uuid",
            "title": "Job Id",
            "type": "string"
          }
        },
        "required": [
          "job_id",
          "created"
        ],
        "title": "LifecycleJobResponse",
        "type": "object"
      },
...
      "ResponseMeta": {
        "description": "Lightweight metadata attached to every response for traceability.",
        "properties": {
          "request_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Request Id"
          },
          "trace_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Trace Id"
          }
        },
        "title": "ResponseMeta",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "AdminBearer": {
        "description": "Deployment admin key for organization management",
        "scheme": "bearer",
        "type": "http"
      },
      "OrganizationApiKey": {
        "description": "Organization API key (`ape_live_\u2026`)",
        "in": "header",
        "name": "X-API-Key",
        "type": "apiKey"
      },
      "OrganizationBearer": {
        "description": "Organization API key (`ape_live_\u2026`)",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "AI Platform Engine (APE) - deployable, provider-agnostic AI infrastructure exposed as Project-scoped REST APIs.",
    "title": "AI Platform Engine",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/projects/{project_id}/documents/{document_id}/index": {
      "post": {
        "operationId": "index_document_api_v1_projects__project_id__documents__document_id__index_post",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "document_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Document Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_DocumentResponse_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OrganizationBearer": []
          },
          {
            "OrganizationApiKey": []
          }
        ],
        "summary": "Enqueue retrieval indexing",
        "tags": [
          "documents"
        ]
      }
    },
...
    "/api/v1/projects/{project_id}/index-builds": {
      "get": {
        "operationId": "list_index_builds_api_v1_projects__project_id__index_builds_get",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_IndexBuildListResponse_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OrganizationBearer": []
          },
          {
            "OrganizationApiKey": []
          }
        ],
        "summary": "List Index Builds",
        "tags": [
          "index-builds"
        ]
      }
    },
    "/api/v1/projects/{project_id}/index-builds/reconcile-storage": {
      "post": {
        "operationId": "reconcile_storage_api_v1_projects__project_id__index_builds_reconcile_storage_post",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_LifecycleJobResponse_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OrganizationBearer": []
          },
          {
            "OrganizationApiKey": []
          }
        ],
        "summary": "Reconcile Storage",
        "tags": [
          "index-builds"
        ]
      }
    },
    "/api/v1/projects/{project_id}/index-builds/reembed": {
      "post": {
        "operationId": "reembed_corpus_api_v1_projects__project_id__index_builds_reembed_post",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_LifecycleJobResponse_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OrganizationBearer": []
          },
          {
            "OrganizationApiKey": []
          }
        ],
        "summary": "Reembed Corpus",
        "tags": [
          "index-builds"
        ]
      }
    },
    "/api/v1/projects/{project_id}/index-builds/reindex": {
      "post": {
        "operationId": "reindex_corpus_api_v1_projects__project_id__index_builds_reindex_post",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_LifecycleJobResponse_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OrganizationBearer": []
          },
          {
            "OrganizationApiKey": []
          }
        ],
        "summary": "Reindex Corpus",
        "tags": [
          "index-builds"
        ]
      }
    },
    "/api/v1/projects/{project_id}/index-builds/rollback": {
      "post": {
        "operationId": "rollback_index_build_api_v1_projects__project_id__index_builds_rollback_post",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_IndexBuildResponse_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OrganizationBearer": []
          },
          {
            "OrganizationApiKey": []
          }
        ],
        "summary": "Rollback Index Build",
        "tags": [
          "index-builds"
        ]
      }
    },
    "/api/v1/projects/{project_id}/index-builds/{build_id}/activate": {
      "post": {
        "operationId": "activate_index_build_api_v1_projects__project_id__index_builds__build_id__activate_post",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "build_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Build Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_IndexBuildResponse_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OrganizationBearer": []
          },
          {
            "OrganizationApiKey": []
          }
        ],
        "summary": "Activate Index Build",
        "tags": [
          "index-builds"
        ]
      }
    },
...
  }
}